Refactor chronicle views, extract collapsible component, and add debug mode
