+++ css/main.css
@@ -367,118 +367,7 @@ li {
 }
 
 /* Entry summary (collapsible) */
-.entry {
-  padding: var(--space-xl) 0;
-}
-
-.entry-header {
-  margin-bottom: var(--space-xl);
-  padding-bottom: var(--space-lg);
-  border-bottom: var(--border-width) solid var(--color-border-light);
-}
-
-.entry-title {
-  margin-bottom: var(--space-lg);
-  text-align: center;
-  padding: 0 var(--space-lg);
-}
-
-.entry-subtitle {
-  margin-bottom: var(--space-lg);
-  padding: 0 var(--space-lg);
-  text-align: center;
-  font-style: italic;
-}
-
-.entry-meta {
-  display: flex;
-  justify-content: space-between;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  font-style: italic;
-}
-
-.entry-actions {
-  display: flex;
-  gap: var(--space-xs);
-}
-
-.entry-timestamp {
-  color: var(--color-text-muted);
-}
-
-.entry-summary {
-  border-bottom: var(--border-width) solid var(--color-border-light);
-  margin-bottom: var(--space-lg);
-}
-
-.entry--placeholder .entry-title h3,
-.entry--placeholder .entry-subtitle p,
-.entry--placeholder .entry-summary p {
-  opacity: 0.5;
-  font-style: italic;
-}
-
-.entry-content--hidden {
-  /* Show up to 2.5 lines, then fade out */
-  max-height: calc(4 * 1.4em); /* 2.5 lines, using line-height-normal */
-  overflow: hidden;
-  position: relative;
-}
-
-.entry-content--hidden::after {
-  content: "";
-  position: absolute;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  height: 2.5em;
-  pointer-events: none;
-  background: linear-gradient(to bottom, rgba(255,255,255,0) 0%, var(--color-background) 95%);
-}
-
-@supports (initial-letter: 2) or (-webkit-initial-letter: 2) {
-  .entry-content p:first-of-type:first-letter {
-    initial-letter: 2;
-    -webkit-initial-letter: 2;
-    margin-right: var(--space-xs);
-    font-style: italic;
-  }
-}
-
-.entry-content p:first-of-type:first-line {
-  text-transform: uppercase;
-  font-style: italic;
-}
-
-.entry-content-controls {
-  display: flex;
-  justify-content: center;
-  margin-top: var(--space-md);
-}
-
-
-.entry-content-control__toggle {
-  background: none;
-  border: none;
-  color: var(--color-text-secondary);
-  font-family: var(--font-body);
-  font-size: var(--font-size-sm);
-  padding: 0;
-  margin: 0;
-  cursor: pointer;
-  /* text-decoration: underline; */
-  font-style: italic;
-  transition: color 0.15s;
-}
-
-.entry-content-control__toggle:hover,
-.entry-content-control__toggle:focus {
-  color: var(--color-accent-hover);
-  outline: none;
-
-}
+/* Entry item styles moved to css/components/entry-item.css */
 
 /* Navigation - Minimal */
 .nav-back {
