+++ js/chronicle.js
@@ -5,68 +5,36 @@ import { ensureChronicleStructure, getChroniclePartsMap } from './chronicle-stat
 import { onJournalChange, onSummariesChange } from './yjs.js';
 import { onChronicleChange } from './yjs.js';
 import { PART_SIZE_DEFAULT, backfillPartsIfMissing, recomputeRecentSummary, recomputePartSummary, recomputeSoFarSummary } from './parts.js';
-import { parseMarkdown } from './utils.js';
 import { formatDate } from './utils.js';
+import { renderSoFar as viewRenderSoFar, renderRecent as viewRenderRecent, renderPartsList as viewRenderPartsList } from './chronicle-views.js';
 
 const renderSoFar = (state) => {
   const el = document.getElementById('so-far-content');
   if (!el) return;
   const soFar = ensureChronicleStructure(state).get('soFarSummary');
-  el.innerHTML = '';
-  const wrapper = createCollapsibleSummary((soFar || 'No summary yet.').trim());
-  el.appendChild(wrapper);
+  viewRenderSoFar(el, soFar);
 };
 
 const renderRecent = (state) => {
   const el = document.getElementById('recent-content');
   if (!el) return;
   const recent = ensureChronicleStructure(state).get('recentSummary');
-  el.innerHTML = '';
-  const wrapper = createCollapsibleSummary((recent || 'No recent summary yet.').trim());
-  el.appendChild(wrapper);
+  viewRenderRecent(el, recent);
 };
 
 const renderPartsList = (state) => {
   const el = document.getElementById('parts-list');
   if (!el) return;
   const latest = ensureChronicleStructure(state).get('latestPartIndex') || 0;
-  if (latest <= 0) {
-    el.textContent = 'No parts yet.';
-    return;
-  }
   const entries = getEntries(state);
   const idToEntry = new Map(entries.map(e => [e.id, e]));
-  const list = document.createElement('div');
+  const parts = getChroniclePartsMap(state);
+  const items = [];
   for (let i = 1; i <= latest; i++) {
-    const partDiv = document.createElement('div');
-    partDiv.className = 'part-list-item';
-
-    const parts = getChroniclePartsMap(state);
     const partObj = parts.get(String(i));
-    const title = (partObj && partObj.get('title')) || `Part ${i}`;
-    const idsJson = JSON.stringify((partObj && partObj.get('entries') && partObj.get('entries').toArray()) || []);
-    let firstTs = '';
-    let lastTs = '';
-    try {
-      const ids = JSON.parse(idsJson || '[]');
-      if (ids.length > 0) {
-        const first = idToEntry.get(ids[0]);
-        const last = idToEntry.get(ids[ids.length - 1]);
-        firstTs = first ? formatDate(first.timestamp) : '';
-        lastTs = last ? formatDate(last.timestamp) : '';
-      }
-    } catch {}
-
-    partDiv.innerHTML = `<div class="part-list-item__header"><strong>${title}</strong>${firstTs && lastTs ? ` — ${firstTs} to ${lastTs}` : ''}</div>`;
-    const viewLink = document.createElement('a');
-    viewLink.href = `/part.html?part=${i}`;
-    viewLink.textContent = 'View Part';
-    viewLink.className = 'btn btn-secondary';
-    partDiv.appendChild(viewLink);
-    list.appendChild(partDiv);
+    items.push({ index: i, title: partObj && partObj.get('title'), entries: partObj && partObj.get('entries') ? partObj.get('entries').toArray() : [] });
   }
-  el.innerHTML = '';
-  el.appendChild(list);
+  viewRenderPartsList(el, items, idToEntry);
 };
 
 const init = async () => {
@@ -121,7 +89,7 @@ const init = async () => {
 
   // Optional debug actions via URL flag ?debug=1
   const debugEnabled = (() => {
-    try { return new URL(window.location.href).searchParams.get('debug') === '1'; } catch { return false; }
+    try { return new URL(window.location.href).searchParams.get('debug') === '1' || localStorage.getItem('debug-mode') === '1'; } catch { return false; }
   })();
   if (debugEnabled) {
     injectDebugControls();
@@ -129,39 +97,6 @@ const init = async () => {
 };
 
 init();
-
-// Helpers
-const createCollapsibleSummary = (text) => {
-  const safeHtml = parseMarkdown(text);
-  const wrapper = document.createElement('div');
-  const toggleButton = document.createElement('button');
-  toggleButton.className = 'entry-summary__toggle';
-  toggleButton.type = 'button';
-  const toggleLabel = document.createElement('span');
-  toggleLabel.className = 'entry-summary__label';
-  toggleLabel.textContent = 'Show';
-  const toggleIcon = document.createElement('span');
-  toggleIcon.className = 'entry-summary__icon';
-  toggleIcon.textContent = '▼';
-  toggleButton.appendChild(toggleLabel);
-  toggleButton.appendChild(toggleIcon);
-
-  const contentDiv = document.createElement('div');
-  contentDiv.className = 'entry-summary__content';
-  contentDiv.style.display = 'none';
-  contentDiv.innerHTML = safeHtml;
-
-  toggleButton.addEventListener('click', () => {
-    const isExpanded = contentDiv.style.display !== 'none';
-    contentDiv.style.display = isExpanded ? 'none' : 'block';
-    toggleButton.classList.toggle('entry-summary__toggle--expanded', !isExpanded);
-    toggleLabel.textContent = isExpanded ? 'Show' : 'Hide';
-  });
-
-  wrapper.appendChild(toggleButton);
-  wrapper.appendChild(contentDiv);
-  return wrapper;
-};
 
 const injectDebugControls = () => {
   const container = document.querySelector('.container-narrow');
